requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==1.11
idna==2.8
Pillow==5.4.1
pkg-resources==0.0.0
python3-memcached==1.51
pytz==2018.9
requests==2.21.0
urllib3==1.24.1
